server/requirements.txt
bottle==0.12.16
CherryPy==8.9.1
mysql-connector-python==8.0.15
